Remove protected_attributes dependency
The protected_attributes gem and strong parameters in Rails 4 don't
play nice together in the same project: protected_attributes
completely takes over the mass assignment sanitization. Since it's
unlikely that transactions will be created from the form parameters
in the first place, this gem does not provide a real benefit.